+++ lib/webidl2.js
@@ -2,6 +2,8 @@
 
 import { argumentNameKeywords, stringTypes, Tokeniser } from "./tokeniser.js";
 import { Base } from "./productions/base.js";
+import { Includes } from "./productions/includes.js";
+import { unescape } from "./helpers.js";
 
 /**
  * @param {Tokeniser} tokeniser
@@ -25,10 +27,6 @@ function parseByTokens(tokeniser) {
 
   function consume(...candidates) {
     return tokeniser.consume(...candidates);
-  }
-
-  function unescape(identifier) {
-    return identifier.startsWith('_') ? identifier.slice(1) : identifier;
   }
 
   function unconsume(position) {
@@ -938,34 +936,6 @@ function parseByTokens(tokeniser) {
     }
   }
 
-  class Includes extends Base {
-    static parse() {
-      const target = consume(ID);
-      if (!target) {
-        return;
-      }
-      const tokens = { target };
-      tokens.includes = consume("includes");
-      if (!tokens.includes) {
-        unconsume(target.index);
-        return;
-      }
-      tokens.mixin = consume(ID) || error("Incomplete includes statement");
-      tokens.termination = consume(";") || error("No terminating ; for includes statement");
-      return new Includes({ source, tokens });
-    }
-
-    get type() {
-      return "includes";
-    }
-    get target() {
-      return unescape(this.tokens.target.value);
-    }
-    get includes() {
-      return unescape(this.tokens.mixin.value);
-    }
-  }
-
   function definition() {
     return callback() ||
       interface_() ||
@@ -973,7 +943,7 @@ function parseByTokens(tokeniser) {
       Dictionary.parse() ||
       Enum.parse() ||
       Typedef.parse() ||
-      Includes.parse() ||
+      Includes.parse(tokeniser) ||
       Namespace.parse();
   }
